Se implementa arquitectura hexagonal al proyecto client-service
