--- client-service/src/main/java/co/com/pragma/ClientServiceApplication.java
@@ -1,15 +1,0 @@
-package co.com.pragma;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@SpringBootApplication
-public class ClientServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ClientServiceApplication.class, args);
-	}
-
-}
