+++ client-service/src/main/java/co/com/pragma/infrastructure/exception/DataNoFoundException.java
@@ -1,0 +1,14 @@
+package co.com.pragma.infrastructure.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@SuppressWarnings("serial")
+@ResponseStatus(code = HttpStatus.FORBIDDEN)
+public class DataNoFoundException extends RuntimeException {
+	
+	public DataNoFoundException() {
+		super("No se encontraron datos");
+	}
+
+}
