+++ client-service/src/main/java/co/com/pragma/infrastructure/exception/ImageTypeIncompatibilityException.java
@@ -1,0 +1,15 @@
+package co.com.pragma.infrastructure.exception;
+
+import co.com.pragma.infrastructure.util.Message;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@SuppressWarnings("serial")
+@ResponseStatus(code = HttpStatus.FORBIDDEN)
+public class ImageTypeIncompatibilityException extends RuntimeException{
+	
+	public ImageTypeIncompatibilityException() {
+		super(Message.IMAGE_TYPE_NOT_FOUND);
+	}
+
+}
