+++ client-service/src/main/java/co/com/pragma/infrastructure/util/Message.java
@@ -1,0 +1,12 @@
+package co.com.pragma.infrastructure.util;
+
+public class Message {
+	
+	public static final String NOT_FOUND_ATRIBUTE = "Los atributos de la entidad no concuerdan con los campos en base de datos";
+	public static final String NULL_DATA = "Esta intentando registrar campos obligatorios con valor null";
+	public static final String DATA_NO_PRESENT = "No se encontraron datos";
+	public static final String DATA_DUPLICATE = "El documento que esta tratando de ingresar ya se encuentra registrado";
+	public static final String IMAGE_TYPE_NOT_FOUND = "El tipo de imagen es incorrecto. Los tipos permitidos son: .PNG, .JPG";
+	public static final String CONTACT_ADMINISTRATOR = "Ocurrio un error por favor contactar al administrador";
+
+}
