+++ client-service/src/test/java/co/com/pragma/ClientEntityServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package co.com.pragma;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ClientEntityServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
