--- client-service/src/test/java/co/com/pragma/ClientServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package co.com.pragma;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ClientServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
